requirements_dash.txt
# Dash Dependencies for ART Call Center Solution
# Base Dash framework
dash>=2.14.0
dash-bootstrap-components>=1.5.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0

# Plotly for charts (already used)
plotly>=5.17.0

# Databricks SDK (existing)
databricks-sdk>=0.1.0

# Data processing (existing)
pandas>=2.0.0
numpy>=1.24.0

# Caching
cachetools>=5.3.0

# Testing
pytest>=7.4.0
pytest-dash>=1.1.0

# Development
python-dotenv>=1.0.0

